Implemented events loading per specific client_id on EventsList screen
